Added Logic2 server
Added python file to interface with the logic analyser. This will be
used to read the input from the deivce. Next step is to host some sort
of interface between the logic output and the python test runner.
Perhaps a simple exchange of data in the python script, or a web server
that is handled by pio core.